+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.5.0 (30-Sep-2020)
+- Try deconstructing xmlDoc instead of traversing the XML
+
+## 0.4.1 (30-Sep-2020)
+- Add more field mappings
+
 ## 0.4.0 (29-Sep-2020)
 - Basic version of Actor5e.js to create an appropriate structure